rajiveutils: Robust Angle Based Joint and Individual Variation Explained

A robust alternative to the aJIVE (angle based Joint and Individual Variation Explained) method (Feng et al 2018: doi:10.1016/j.jmva.2018.03.008 ) for the estimation of joint and individual components in the presence of outliers in multi-source data. It decomposes the multi-source data into joint, individual and residual (noise) contributions. The decomposition is robust to outliers and noise in the data. The method is illustrated in Ponzi et al (2021) arXiv:2101.09110.

See also

Author

Maintainer: Erica Ponzi erica.ponzi@medisin.uio.no

Authors:

Abhik Ghosh abhik.ghosh.stat@gmail.com